perf(components): eliminate barrel files for tree-shaking optimization
Removed barrel files (index.ts) from components/icons and components/skeletons to prevent unnecessary React components from being pulled into the module graph. Separated ICON_IDS into a standalone module and added sideEffects: false to package.json to enable aggressive dead-code elimination.
