+++ package.json
@@ -2,6 +2,7 @@
   "name": "stellarflow-frontend",
   "version": "0.1.0",
   "private": true,
+  "sideEffects": false,
   "scripts": {
     "dev": "next dev",
     "build": "next build && node scripts/check-bundle-size.js",
