+++ src/app/DashboardInteractive.tsx
@@ -3,13 +3,11 @@
 import React, { useEffect, useState } from "react";
 import dynamic from "next/dynamic";
 import { motion, AnimatePresence } from "framer-motion";
-import {
-  Shimmer,
-  MapSkeleton,
-  RateSparklineSkeleton,
-  PriceFeedCardSkeleton,
-  DashboardTrafficChartSkeleton,
-} from "@/components/skeletons";
+import { Shimmer } from "@/components/skeletons/Shimmer";
+import { MapSkeleton } from "@/components/skeletons/MapSkeleton";
+import { RateSparklineSkeleton } from "@/components/skeletons/RateSparklineSkeleton";
+import { PriceFeedCardSkeleton } from "@/components/skeletons/PriceFeedCardSkeleton";
+import { DashboardTrafficChartSkeleton } from "@/components/skeletons/DashboardTrafficChartSkeleton";
 import { useMounted } from "@/app/hooks/useMounted";
 import WebSocketTest from "./components/test/WebSocketTest";
