+++ src/app/relayers/page.tsx
@@ -15,7 +15,8 @@ import {
   Signal,
 } from 'lucide-react';
 import { RELAYERS_PAGE_STATUS_VARIANTS } from '@/lib/classNameVariants';
-import { Icon, ICON_IDS } from "@/components/icons";
+import Icon from "@/components/icons/Icon";
+import { ICON_IDS } from "@/components/icons/iconIds";
 import {
   RelayerManagementRow,
   type RelayerManagementRecord,
