--- src/components/icons/index.ts
@@ -1,4 +1,0 @@
-export { default as SvgSprite } from "./SvgSprite";
-export { ICON_IDS } from "./SvgSprite";
-export type { IconId } from "./SvgSprite";
-export { default as Icon } from "./Icon";
