--- src/components/skeletons/index.ts
@@ -1,10 +1,0 @@
-export { Shimmer } from './Shimmer';
-export { MetricCardSkeleton } from './MetricCardSkeleton';
-export { MapSkeleton } from './MapSkeleton';
-export { RateSparklineSkeleton } from './RateSparklineSkeleton';
-export { PriceFeedCardSkeleton } from './PriceFeedCardSkeleton';
-export { DashboardTrafficChartSkeleton } from './DashboardTrafficChartSkeleton';
-export { StatsSkeleton } from './StatsSkeleton';
-export { AdminStatsSkeleton } from './AdminStatsSkeleton';
-export { ValidatorMetricsSkeleton } from './ValidatorMetricsSkeleton';
-export { ValidatorListSkeleton } from './ValidatorListSkeleton';
